% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.profileLadder.R
\name{as.profileLadder}
\alias{as.profileLadder}
\title{S3 Method class \code{profileLadder}}
\usage{
as.profileLadder(x)
}
\arguments{
\item{x}{an object of the class \code{matrix} or \code{triangle}}
}
\value{
an object of the class \code{profileLadder} -- a list with the following elements:
\item{reserve}{basic summary of the run-off triangle and the true and the estimated reserve
(if available -- \code{NA} values returned otherwise)}
\item{method}{type of the triangle (run-off/complete) or the type of the estimation method used
to complete the functional development profiles (PARALLAX, REACT, or MACRAME)}
\item{completedProfiles}{completed development profiles estimated by the PARALLAX, REACT, or MACRAME algorithm}
\item{inputTriangle}{standard (triangular shaped)  run-off triangle}
\item{trueComplete}{true completed development profiles of teh run-off triangle (if available) -- value \code{NA} returned otherwise}
}
\description{
A function to make the work with the functional development profiles within run-off triangles
more easy and straightforward (particularly when plotting the functional profiles)
}
\examples{
data(CameronMutual)
print(CameronMutual) 

x <- as.profileLadder(CameronMutual)

print(x)
plot(x)

}
\seealso{
\code{\link[=parallelReserve]{parallelReserve()}}, \code{\link[=mcReserve]{mcReserve()}}, \code{\link[=permuteReserve]{permuteReserve()}}, \code{\link[=plot.profileLadder]{plot.profileLadder()}}
}
