% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFBI.R
\docType{data}
\name{CFBI}
\alias{CFBI}
\title{CFBI data}
\format{
\subsection{CFBI}{

A simple \code{10x10} matrix of a class \code{triangle} with ten origin years (rows) each being fully developed
within ten consecutive development periods/years (columns)

\describe{
\item{origin}{matrix rows with the occurrence year (origin)}
\item{dev}{matrix columns with the development period (development)}
}
}
}
\source{
\url{https://www.casact.org/research/reserve_data/othliab_pos.csv}
(NAIC group code: 13641)
}
\usage{
data(CFBI)
}
\description{
An illustrative dataset---a matrix (of the dimensions 10\code{x}10) with ten completed years of claims payment developments of
the Colorado Farm Bureau Insurance company from the period 1988 -- 1997.  The data matrix contains ten origin/occurrence
years (in rows where the first row represents the incident year 1988) with ten consecutive development periods/years (in columns).
}
\details{
The incremental form of the run-off triangle (the upper-left triangular part of the data matrix) contains zeros and negative
incremental payments, therefore, thi run-off triangle can not be used for a standard over-dispersed Poisson model (GLM regression).

In practice, the upper-left triangle (the run-off triangle) is typically observed (known) while the bottom-right
triangular part of the data matrix is treated as a future payments outcome (an "unknown" truth) that should be estimated/predicted.
The Colorado Farm Bureau Insurance data matrix is fully observed to allow for some goodness-of-fit evaluations
}
\references{
Meyers, G. G. and P. Shi (2011). Loss reserving data pulled from NAIC Schedule P. Available from
\url{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}
}
