% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CZ.casco.R
\docType{data}
\name{CZ.casco}
\alias{CZ.casco}
\title{Vehicles damages within a conventional full Casco insurance}
\format{
\subsection{CZ.casco}{

Two run-off triangles (objects of the class \code{triangle}) with the dimensions
\code{17x17} with \code{17} origin years (from the period 2003 -- 2019) and \code{17}
development periods (years again).

\describe{
\item{origin}{matrix rows with the occurrence year (origin)}
\item{dev}{matrix columns with the development period (development)}
}
}
}
\source{
Annonymous major insurance company in the Czech Republic
}
\usage{
data(CZ.casco)
}
\description{
Illustrative dataset provided by the major insurance company in the Czech Republic.
The dataset contains private and commercial business sold to policy holders based
in the Czech Republic and covers the damages to the vehicles within
a framework of the conventional full Casco insurance.
}
\details{
The data are structured in the the list object \code{CZ.casco} with two
elements---the run-off triangle \code{GrossPaid} containing cumulative amounts
of gross paid claims, and another cummulative run-off triangle \code{RBNS}
providing the amounts of the RNNS reserve. The amounts of given in thousands
of the Czech crows (CZK).
}
