% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.R
\name{observed}
\alias{observed}
\title{Observed run-off triangle layout vs. predicted (unknown) layout}
\usage{
observed(object, cum = T)
}
\arguments{
\item{object}{either an integer value to denote the dimension of the run-off triangle (i.e.,
the number of origins (rows) and the development periods (columns)) in the output layout,
alternatively, cumulatige or incremental run-off triangle  which is an object of the
class \code{matrix} or \code{triangle}}

\item{cum}{logical to indicate whether the output run-off triangle is supposed to
be of a cumulative type (\code{DEFAULT}) or an incremental type (\code{cum = FALSe}).
If the input \code{object} is an integer value (i.e., the dimension of the run-off triangle),
than the choice of the \code{cum} paramter is ignored}
}
\value{
If \code{object} is an integer value then the function returns a TRUE/FALSE layout matrix
with \code{TRUE} values for the observed (known) part of the matrix (the upper-left triangular part of the matrix -- the run-off triangle)
and values \code{FALSE} otherwise.  If \code{object} is a matrix (an object of the type \code{matrix} or \code{triangle})
then the function returns the observed (known) part of the run-off triangle with \code{NA} values elsewhere. Depending
on the choice of the \code{cum} paramter, either cumulative (\code{DEFAULT}) or incremental (\code{cum = FALSE}) run-off
triangle is returned
}
\description{
Simple layout function to make work with (cumulative or incremental) run-off triangles
more easy and straightforward.
}
\examples{
print(observed(5))
print(!observed(5))

data(CameronMutual)
observed(CameronMutual)
observed(CameronMutual, cum = FALSE)

}
\seealso{
\code{\link[=plot.profileLadder]{plot.profileLadder()}}, \code{\link[=parallelReserve]{parallelReserve()}}, \code{\link[=mcReserve]{mcReserve()}}
}
