% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuteReserve.R
\name{permuteReserve}
\alias{permuteReserve}
\title{Permutation bootstrap reserve (PARALLAX, REACT, MACRAME)}
\usage{
permuteReserve(object, B = 500, std = TRUE, quantile = 0.995)
}
\arguments{
\item{object}{an object of the class \code{profileLadder} (i.e., the output from the function \code{parallelReserve()} or \code{mcReserve()})}

\item{B}{of permutations to be performed (\code{DEFAULT = 500})}

\item{std}{logical to indicate whether the run-off triangle should be standardized by the first column increments (DEFAULT) or not (\code{std = FALSE})}

\item{quantile}{...(DEFAULT \code{quantile = 0.995})}
}
\value{
A list with with the following elements:
\item{method}{summary of the permutation bootstrap}
\item{estimatedReserve}{numeric vector with four values: Total paid amount (i.e., the sum of the last observed diagonal in a cumulative run-off triangle); Total estimated amount (i.e., the sum of the last column
in the completed cumulative triangle); Estimated reserve (i.e., the sum of the last column in the completed cumulative triangle minus the sum of the last observed diagonal); True reserve
if a completed (true) run-off triangle is available}
\item{pSummary}{numeric vector with our values: Average of \code{B} permutation bootstrap reserves; Standard error of \code{B} permutation bootstrap reserves;
numeric value \code{BootCov\%} given as a percentage proportion between the standard error and the average; \code{BootVar.995} providing the 0.995 quantile of the permutation bootstrap reserves
relative to the permutation bootstrap mean reserve}
\item{pReserves}{a numeric vector of the length \code{B} with estimated (permuted) reserves for each row-permuted run-off triangle}
\item{time}{a numeric vector with the number of permutation resamples (\code{B}), the run-off triangle dimension (\code{n}) and
the computation time consumed for the permutation bootstrap calculations}
}
\description{
The function takes the output from the function \code{parallelReserve()} or \code{mcReserve} and provides the overall
reserve distribution in terms of the permutation bootstrap proposed in Maciak, Mizera, and Pešta (2022).
}
\examples{
## REACT algorithm and the permutation bootstrap reserve 
data(CameronMutual)

output <- parallelReserve(CameronMutual, method = "react")
permuteReserve(output)

## MACRAME algorithm with a pre-specified number of states and the permutation bootstrap reserve
output <- mcReserve(CameronMutual, states = 5)
permuteReserve(output)


}
\references{
Maciak, M., Mizera, I., and Pešta, M. (2022). Functional Profile Techniques for Claims Reserving. ASTIN Bulletin, 52(2), 449-482. DOI:10.1017/asb.2022.4
}
\seealso{
\code{\link[=parallelReserve]{parallelReserve()}}, \code{\link[=mcReserve]{mcReserve()}}
}
