% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.permutedReserve.R
\name{plot.permutedReserve}
\alias{plot.permutedReserve}
\title{Plotting the output of the permutation bootstrap}
\usage{
\method{plot}{permutedReserve}(x, density = FALSE, ...)
}
\arguments{
\item{x}{an object of the class \code{permutedReserves} -- the output form the \code{permuteReserve()} function}

\item{density}{logical to indicate whether a simple boxplot is supposed to be plotted (DEFAULT) or more complex histogram with a nonparametrix density estimate (\code{density = TRUE})}

\item{...}{other graphical parameters to plot}
}
\value{
The function returns either the boxplot of the permuted bootstrap reserves (\code{density = FALSE}) or it
returns a histogram with a nonparametric estimate of the overall (permuted bootstrapped) reserve (\code{density = TRUE}).
Some additional quantities (such as the true reserve, or the estimated reserve are provided in the plot)
}
\description{
The function provides a graphical visualization of the results obtained from the permutation
bootstrap (see Maciak, Mizera, and Pesta (2022) for further details) applied to the output of
the \code{parallelReserve()} function or the \code{mcReserve()} function.
}
\examples{
## reserve estimate by MACRAME and boxplot of permuted reserves
x <- mcReserve(CameronMutual)
plot(permuteReserve(x))

## histogram of and the estimated distributional reserve
plot(permuteReserve(x), density = TRUE)

}
\seealso{
\code{\link[=permuteReserve]{permuteReserve()}}
}
