% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.profileLadder.R
\name{plot.profileLadder}
\alias{plot.profileLadder}
\title{Plotting development profiles}
\usage{
\method{plot}{profileLadder}(x, xlab = "Development Year", ylab = "Cumulative Claims", main = "", ...)
}
\arguments{
\item{x}{an object of the class \code{profileLadder} -- the output form \code{parallelReserve()}  or \code{mcReserve()}}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}

\item{main}{title of the plot}

\item{...}{other graphical parameters to plot}
}
\value{
A graph with the observed functional development profiles from the input run-off triangle,
predicted functional segments (functional profile completions) by the corresponding estimation method (PARALLAX, REACT, or MACRAME)
and true (i.e., unknown future) profiles (if available)
}
\description{
The function provides a graphical representation of the completed functional profiles provided by the PARALLAX, REACT, or MACRAME
algorithm (see Maciak, Mizera, and Pesta (2022) for further details). The function takes an object of class 'profileLadder' --  the
output of the \code{parallelReserve()} function or the \code{mcReserve()} function.
}
\examples{
## run-off triangle with 'unknown' (future) payments
data(CFBI)
print(triangle <- CFBI)
plot(mcReserve(triangle))

## standard run-off triangle (with unknown future)
print(triangle <- observed(CFBI))
plot(mcReserve(triangle))


}
\seealso{
\code{\link[=observed]{observed()}}, \code{\link[=parallelReserve]{parallelReserve()}}, \code{\link[=mcReserve]{mcReserve()}}
}
