% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mcSetup.R
\name{print.mcSetup}
\alias{print.mcSetup}
\title{Print Objects of the S3 Class \code{mcSetup}}
\usage{
\method{print}{mcSetup}(x, ...)
}
\arguments{
\item{x}{an object of the class \code{mcSetup} resulting from a call  of the
\code{incrExplor()} function}

\item{...}{further arguments passed to \code{print}}
}
\value{
displays information resulting from a call of the \code{incrExplor()}
function
}
\description{
Function to organize and print the output provided by the function \code{incrExplor()}
}
\examples{
data(CameronMutual)
x <- incrExplor(CameronMutual)
print(x) 

}
\seealso{
\code{\link[=incrExplor]{incrExplor()}}, \code{\link[=mcReserve]{mcReserve()}}, \code{\link[=mcBreaks]{mcBreaks()}}, \code{\link[=mcStates]{mcStates()}}
}
