% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.profileLadder.R
\name{print.profileLadder}
\alias{print.profileLadder}
\title{Print objects of S3 class \code{profileLadder}}
\usage{
\method{print}{profileLadder}(x, ...)
}
\arguments{
\item{x}{an object of the class \code{profileLadder} resulting from a call to the
\code{parallelReserve()} or \code{mcReserve} function}

\item{...}{further arguments passed to printfurther arguments passed to \code{print}}
}
\value{
an object of the class \code{profileLadder} -- a list with the following elements:
}
\description{
Function to print the results of a call to the \code{parallelReserve()} and \code{mcReserve} functions.
}
\examples{
data(CameronMutual)
x <- as.profileLadder(CameronMutual)
print(x) 

}
\seealso{
\code{\link[=as.profileLadder]{as.profileLadder()}}, \code{\link[=parallelReserve]{parallelReserve()}}, \code{\link[=mcReserve]{mcReserve()}}
}
