% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.profileLadder.R
\name{summary.profileLadder}
\alias{summary.profileLadder}
\title{Summary method for an object of the S3 class method \code{profileLadder}}
\usage{
\method{summary}{profileLadder}(object, plotOption = FALSE, ...)
}
\arguments{
\item{object}{an object of a class 'triangle', 'matrix', or 'profileLadder' -- either a run-off triangle or the output
form the \code{parallelReserve()} or \code{mcReserve()} functions}

\item{plotOption}{logical to indicate whether a graphical output shouls be provided (set to \code{FALSE} by DEFAULT). If the  residuals
(standard or back-fitted) are provided in \code{x} the plot summarizes the residuals otherwise a simple barplot is
given to summarize the estimated reserve}

\item{...}{not used}
}
\value{
Summary of the completed functional profiles and the estimated reserve (provided by the
function \code{parallelReserve()} or \code{mcReserve()}). Summary of the incremental residuals
(standard or backfitted) is also provided if the residuals  are available.
The output is a list with the following items:
\item{origins}{a matrix with the row-specific summary of the completed functional profiles (except the first
fully developed profile -- the first row in the run-off triangle).
The first column of the matrix (\code{First}) gives the first origin payments; The second column (\code{Latest}) gives
the last available (cumulative) payments (i.e., values from the last running diagonal in the run-off triangle);
The third column (\code{Dev.To.Date}) give a relative proportion of the paid amount (\code{Latest}) with respect to the
estimated ultimate (\code{Ultimate}) given in the fourth column; Finally, the last column (\code{IBNR}) gives the
estimated amount still left to pay.}
\item{overall}{Summary of the estimated reserve; TLatest paid amount (sum of the last running diagonal);
Estimated reserve; True reserv (if available); Accuracy in terms of \code{Reserve\%} which is obtained
as a percentage of the estimated reserve with respect to the true reserve (see Maciak, Mizera, and Pešta (2022)
for more details) and, finally, the overall \code{Dev.To.Date} value.}
\item{resids}{Basic description characteristics of the residuals (standard or back-fitted) if the residuals are provided in \code{x}}
}
\description{
The function provides a brief summary of the output from the functions parallelReserve() and mcReserve().
}
\examples{
data(CameronMutual)
summary(CameronMutual)

## standard summary output
summary(mcReserve(CameronMutual))

## summary output with plotOption = TRUE
summary(mcReserve(CameronMutual), plotOption = TRUE)

## summary output with (standard) residuals and plotOption = TRUE
summary(mcReserve(CameronMutual, residuals = TRUE), plotOption = TRUE)

#' ## summary output with (back-fitted) residuals and plotOption = TRUE
summary(mcReserve(observed(CameronMutual), residuals = TRUE), plotOption = TRUE)


}
\seealso{
\code{\link[=as.profileLadder]{as.profileLadder()}}, \code{\link[=parallelReserve]{parallelReserve()}}, \code{\link[=mcReserve]{mcReserve()}}
}
