% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xNetSubscribe.R
\docType{data}
\name{xNetSubscribe}
\alias{xNetSubscribe}
\title{Data provider income}
\format{
\subsection{xNetSubscribe}{

A simple \code{12x12} (trangular) matrix of a class \code{triangle} with 12 months when new customers subscribed
to the data stream service (rows) and consecutive montly-based payments (columns).
}
}
\usage{
data(xNetSubscribe)
}
\description{
An illustrative dataset---a matrix (of the dimensions 12\code{x}12) with a monthly-based income (in EUR)
of a local internet data provider with the income structured by the customers subscribing
within the given month (in 2023) given in rows and monthly-based payments in 2023 in columns.
}
