% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.profileLadder.R
\name{print.profileLadder}
\alias{print.profileLadder}
\title{Print objects of the S3 class \code{profileLadder}}
\usage{
\method{print}{profileLadder}(x, ...)
}
\arguments{
\item{x}{an object of the class \code{profileLadder} resulting from a call
of one of the functions \code{parallelReserve()}, \code{mcReserve}, or
\code{as.profileLadder()}}

\item{...}{further arguments passed to \code{print}}
}
\value{
displays information resulting from a call of the \code{parallelReserve()}
function or the \code{mcReserve} function
}
\description{
Function to organize and print the outputs provided by the function
\code{parallelReserve()} and the function \code{mcReserve}
}
\examples{
data(CameronMutual)
x <- as.profileLadder(CameronMutual)
print(x) 

}
\seealso{
\code{\link[=as.profileLadder]{as.profileLadder()}}, \code{\link[=parallelReserve]{parallelReserve()}}, \code{\link[=mcReserve]{mcReserve()}}
}
