% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xNetSubscribe.R
\docType{data}
\name{xNetSubscribe}
\alias{xNetSubscribe}
\title{Data provider monthly income}
\format{
\subsection{xNetSubscribe}{

A simple \code{12x12} (trangular) matrix of the class \code{triangle} with
twelve consecutive months (January 2023 -- December 2023) when new customers
subscribed to the stream service (rows) and monthly-based payments (columns)
}
}
\usage{
data(xNetSubscribe)
}
\description{
An illustrative dataset---a matrix (of the dimensions 12\code{x}12) with a
monthly-based income (in EUR) of a local internet data provider with the
income structured by the customers subscribing  within the given month (in 2023)
reported in the rows and monthly-based payments reported in columns.
The data matrix represents the incremental type of the run-off triangle.
}
